Moved token extraction in @after section in RegistrationTest.java, unnecessary spaces removed in ConstructorTest.java, LoginTest.java, RegistrationTest.java,TransitionTest.java

+++ src/test/java/RegistrationTest.java
@@ -19,7 +19,8 @@ public class RegistrationTest {
     private UserClient userClient;
     private String email = "burger-test1@yandex.ru";
     private String password = "test12";
-    private String token;
+    private String token = "";
+    private String rawToken = "";
 
     //Чтобы запустить тесты в FireFox нужно в setDriver() поменять ChromeDriver() на FireFoxDriver()
     @Before
@@ -32,48 +33,41 @@ public class RegistrationTest {
 
     @After
     public void teardown() {
-        driver.quit();
+        ValidatableResponse response = userClient.login(new UserCredentials(email, password));
+        int responseCode = response.extract().statusCode();
+        if (responseCode == 200) {
+            rawToken = response.extract().path("accessToken");
+            token = rawToken.replaceFirst("Bearer ", "");
+            userClient.delete(token);
+        }
+            driver.quit();
     }
 
     @Test
     @DisplayName("Registration test")
     public void registrationTest() {
         RegisterPage registerPage = new RegisterPage(driver);
-
         registerPage.enterName("burger-test");
         registerPage.enterEmail(email);
         registerPage.enterPassword(password);
         registerPage.clickEnterInPasswordField();
         registerPage.clickRegisterButton();
-
         LoginPage loginPage = new LoginPage(driver);
-
         new WebDriverWait(driver, 5)
                 .until(ExpectedConditions.visibilityOfElementLocated(loginPage.getLoginPageHeader()));
-
         String currentUrl = driver.getCurrentUrl();
         String expectedUrl = "https://stellarburgers.nomoreparties.site/login";
-
         assertEquals("Wrong url", currentUrl, expectedUrl);
-
-        ValidatableResponse response = userClient.login(new UserCredentials(email, password));
-        String rawToken = response.extract().path("accessToken");
-
-        token = rawToken.replaceFirst("Bearer ", "");
-
-        userClient.delete(token);
     }
 
     @Test
     @DisplayName("Registration test with short password")
     public void registrationWithShortPasswordTest() {
         RegisterPage registerPage = new RegisterPage(driver);
-
         registerPage.enterName("burger-test");
         registerPage.enterEmail(email);
         registerPage.enterPassword("test1");
         registerPage.clickRegisterButton();
-
         Assert.assertTrue(registerPage.incorrectPasswordErrorIsVisible());
     }
